trim hybridation against Python < 3.5
