+++ behave/log_config.py
@@ -93,7 +93,7 @@ class LoggingConfigurator:
     def make_formatter(self, format=None, datefmt=None):
         format_ = format or self.config.logging_format
         datefmt = datefmt or self.config.logging_datefmt
-        if format_ and _PYTHON_VERSION >= (3, 2):
+        if format_:
             # -- SINCE: Python 3.2 -- logging.Formatter(..., style)
             style = self.select_format_style(format_)
             return logging.Formatter(format_, datefmt, style)
