+++ tests/unit/test_deprecated.py
@@ -4,6 +4,10 @@ Unit test facade to protect pytest runner from 3.5 grammar changes:
 * Runs tests related to async-steps.
 """
 
+from behave.api.async_step import (
+    async_run_until_complete,
+    use_or_create_event_loop,
+)
 from behave.capture import CaptureController
 from behave.configuration import Configuration
 from behave.python_feature import PythonFeature
@@ -14,16 +18,26 @@ import pytest
 
 
 # -----------------------------------------------------------------------------
-# TEST SUITE -- PYTHON >= 3.5
-# -----------------------------------------------------------------------------
-if PythonFeature.has_async_function():
-    # -- REQUIRES: async/await syntax and coroutine support (since: Python 3.5).
-    from ._test_deprecated_py35 import *  # noqa: F403
-
-
-# -----------------------------------------------------------------------------
 # TEST SUITE -- DEPRECATED THINGS
 # -----------------------------------------------------------------------------
+
+class TestApiAsyncStep:
+    """
+    Covers :mod:`behave.api.async_step`.
+    """
+
+    def test_deprecated_async_run_until_complete(self):
+        """Ensure that the ``@async_run_until_complete`` issues a warning."""
+        with pytest.warns(PendingDeprecationWarning, match=r"@async_run_until_complete.*"):
+            @async_run_until_complete
+            async def async_step_example(ctx):
+                pass
+
+    def test_deprecated_use_or_create_event_loop(self):
+        with pytest.warns(PendingDeprecationWarning, match=r"use_or_create_event_loop:.*"):
+            _ = use_or_create_event_loop()
+
+
 class TestActiveTagMatcher:
     def test_deprecated_should_exclude_with(self):
         active_tag_matcher = ActiveTagMatcher({})
